hoai2025: freeplay updates
