+++ apps/src/bubbleChoice/BubbleChoice.tsx
@@ -11,7 +11,11 @@ import React, {useEffect, useMemo, useRef} from 'react';
 
 import {levelById} from '@cdo/apps/code-studio/progressReduxSelectors';
 import continueOrFinishLesson from '@cdo/apps/lab2/progress/continueOrFinishLesson';
-import {BubbleChoiceLevelData, LabProps} from '@cdo/apps/lab2/types';
+import {
+  BubbleChoiceLevelData,
+  BubbleChoiceSublevel,
+  LabProps,
+} from '@cdo/apps/lab2/types';
 import EnhancedSafeMarkdown from '@cdo/apps/templates/EnhancedSafeMarkdown';
 import ProgressBubble from '@cdo/apps/templates/progress/ProgressBubble';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
@@ -20,11 +24,11 @@ import {
   LevelStatus,
 } from '@cdo/generated-scripts/sharedConstants';
 
+import {navigateToLevelId} from '../code-studio/progressRedux';
 import {commonI18n} from '../types/locale';
 
 import MusicDanceAi from './customModes/MusicDanceAi';
 import {BubbleChoiceLevelProperties} from './types';
-import useNavigateToSublevel from './useNavigateToSublevel';
 
 import styles from './BubbleChoice.module.scss';
 
@@ -53,6 +57,10 @@ const BubbleChoice: React.FC<LabProps<BubbleChoiceLevelProperties>> = ({
           sublevel.level_id
         )?.status || LevelStatus.not_tried
     )
+  );
+
+  const currentLessonId = useAppSelector(
+    state => state.progress.currentLessonId
   );
 
   const [containerWidth, setContainerWidth] = React.useState(0);
@@ -125,7 +133,13 @@ const BubbleChoice: React.FC<LabProps<BubbleChoiceLevelProperties>> = ({
     return level;
   };
 
-  const navigateToSublevel = useNavigateToSublevel();
+  const navigateToSublevel = (sublevel: BubbleChoiceSublevel) => {
+    if (currentLessonId) {
+      dispatch(navigateToLevelId(sublevel.level_id));
+    } else {
+      window.location.href = sublevel.url;
+    }
+  };
 
   if (
     channel &&
@@ -169,7 +183,7 @@ const BubbleChoice: React.FC<LabProps<BubbleChoiceLevelProperties>> = ({
                 width: imageWidth,
                 height: imageWidth / aspectRatio,
               }}
-              onClick={() => navigateToSublevel(levelProperties, sublevel)}
+              onClick={() => navigateToSublevel(sublevel)}
             >
               <div
                 className={styles.sublevelImageContainer}
